Add separate sidebaritems file to not import from enterprise
This has the positive side-effect that things that aren't available in the OSS version won't show in the site-admin UI anymore. They were dead links that linked to a 404 page anyways.
